Check for existance of file before replacing

+++ resources/lib/diagnostics/play_stats.py
@@ -130,7 +130,7 @@ class PlayStatistics:
             save_path = path + '.old'
             directory = os.path.dirname(path)
             os.makedirs(directory, mode=0o751, exist_ok=True)
-            if clz._first_call:
+            if clz._first_call and save_path.exists():
                 os.replace(path, save_path)
                 clz.first_call = False
